Revert "Fix ClangTidy bugprone-integer-division causing failed build"
This reverts commit 315c1711d1e3b1810d30d9d3b154d1eabf3930be.

+++ wpimath/src/main/native/include/frc/MathUtil.h
@@ -332,7 +332,7 @@ constexpr Translation2d SlewRateLimit(const Translation2d& current,
   }
   if (dist > maxVelocity * dt) {
     // Move maximum allowed amount in direction of the difference
-    return current + diff * (static_cast<double>(maxVelocity * dt / dist));
+    return current + diff * (maxVelocity * dt / dist);
   }
   return next;
 }
@@ -362,7 +362,7 @@ constexpr Translation3d SlewRateLimit(const Translation3d& current,
   }
   if (dist > maxVelocity * dt) {
     // Move maximum allowed amount in direction of the difference
-    return current + diff * (static_cast<double>(maxVelocity * dt / dist));
+    return current + diff * (maxVelocity * dt / dist);
   }
   return next;
 }
